refactor(h2r): move parseAndAdapt function to its own file
