+++ src/shared/components/HtmlToReactConverter/parseAndAdapt/filterAttributes.js
