+++ src/shared/components/HtmlToReactConverter/parseAndAdapt/htmlMap.js
