+++ src/shared/components/HtmlToReactConverter/parseAndAdapt/svgMap.js
